requirements-dev.txt
pytest>=7.4
typer>=0.9
Pillow>=10.0
PyMuPDF>=1.23
faster-whisper>=0.10
